# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema TKD API
  version: 1.0.0
  description: Contrato del sistema. PostgreSQL es la fuente de verdad y la escuela se obtiene exclusivamente de la sesión.
servers:
  - url: /api
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tkd_session
  parameters:
    Csrf:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 20 }
    Idempotency:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Conflict:
      description: La versión cambió o la operación contradice el estado actual.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
...
    PaymentWrite:
      type: object
      additionalProperties: false
      required: [chargeId,amount,date]
      properties:
        chargeId: { type: string, format: uuid }
        amount: { type: string, pattern: '^\\d+(\\.\\d{1,2})?$' }
        date: { type: string, format: date }
        method: { type: string, enum: ['',Efectivo,Transferencia,Tarjeta,Otro] }
        observation: { type: string }
        reason: { type: string, description: Obligatorio al corregir un pago existente. }
        version: { type: integer, minimum: 1 }
    VoidPaymentWrite:
      type: object
      required: [reason,version]
      properties: { reason: { type: string, minLength: 1 }, version: { type: integer, minimum: 1 } }
    DiscountWrite:
      type: object
      additionalProperties: false
      required: [studentId,kind,value,reason,recurring]
      properties:
        studentId: { type: string, format: uuid }
        kind: { type: string, enum: [fixed,percentage] }
        value: { type: integer, minimum: 1, description: Centavos si es fijo; puntos base si es porcentaje (25 % = 2500). }
        reason: { type: string, minLength: 1 }
        period: { type: string, pattern: '^\d{4}-\d{2}$' }
        recurring: { type: boolean }
    OrderItemWrite:
      type: object
      additionalProperties: false
      required: [name,quantity,price]
      properties:
        name: { type: string, minLength: 1 }
        size: { type: string }
        quantity: { type: integer, minimum: 1 }
        price: { type: string, pattern: '^\\d+(\\.\\d{1,2})?$' }
    OrderWrite:
      type: object
      additionalProperties: false
      required: [studentId,dueDate,items]
      properties:
        studentId: { type: string, format: uuid }
        dueDate: { type: string, format: date }
        items: { type: array, minItems: 1, items: { $ref: '#/components/schemas/OrderItemWrite' } }
...
    OrderBulkProgressItem:
      type: object
      additionalProperties: false
      required: [id,version]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
    OrderBulkProgressWrite:
      type: object
      additionalProperties: false
      required: [mode,items]
      properties:
        mode: { type: string, enum: [receivePending,deliverReceived] }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/OrderBulkProgressItem' }
paths:
...
  /payments:
    post: { operationId: createPayment, parameters: [{ $ref: '#/components/parameters/Csrf' },{ $ref: '#/components/parameters/Idempotency' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentWrite' } } } }, responses: { '201': { description: Pago registrado }, '200': { description: Reintento idempotente }, '400': { description: Importe inválido o sobrepago } } }
  /payments/{id}/void:
    post: { operationId: voidPayment, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VoidPaymentWrite' } } } }, responses: { '204': { description: Pago anulado con historial } } }
  /payments/{id}:
    put: { operationId: correctPayment, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentWrite' } } } }, responses: { '204': { description: Pago corregido y cambio conservado en historial }, '409': { $ref: '#/components/responses/Conflict' } } }
  /discounts:
    post: { operationId: createDiscount, parameters: [{ $ref: '#/components/parameters/Csrf' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DiscountWrite' } } } }, responses: { '201': { description: Descuento futuro registrado }, '400': { description: Datos o alumno inválidos } } }
  /orders:
    post: { operationId: createOrder, parameters: [{ $ref: '#/components/parameters/Csrf' },{ $ref: '#/components/parameters/Idempotency' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrderWrite' } } } }, responses: { '201': { description: Pedido y cobros por artículo creados atómicamente }, '200': { description: Reintento idempotente del alta } } }
  /orders/{id}/progress:
    post: { operationId: updateOrderProgressBulk, parameters: [{ $ref: '#/components/parameters/ResourceId' },{ $ref: '#/components/parameters/Csrf' },{ $ref: '#/components/parameters/Idempotency' }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrderBulkProgressWrite' } } } }, responses: { '204': { description: Todos los artículos activos avanzaron atómicamente o el reintento ya había sido aplicado }, '400': { description: Modo, lista o cantidades inválidas }, '404': { description: Pedido no encontrado en la escuela }, '409': { $ref: '#/components/responses/Conflict' } } }
...